# apps/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kaneo API",
    "description": "Kaneo Project Management API - Manage projects, tasks, labels, and more",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://cloud.kaneo.app/api",
      "description": "Kaneo API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication"
      },
      "apiKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "apiKeyCookie",
        "description": "API Key authentication via cookie"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/project": {
      "get": {
        "operationId": "listProjects",
        "tags": [
          "Projects"
        ],
        "description": "Get all projects in a workspace",
        "responses": {
          "200": {
            "description": "List of projects with statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "workspaceId": {
                        "type": "string"
                      },
                      "slug": {
                        "type": "string"
                      },
                      "icon": {
                        "type": "string",
                        "nullable": true
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string",
                        "nullable": true
                      },
                      "createdAt": {},
                      "isPublic": {
                        "type": "boolean",
                        "nullable": true
                      }
                    },
                    "required": [
                      "id",
                      "workspaceId",
                      "slug",
                      "icon",
                      "name",
                      "description",
                      "createdAt",
                      "isPublic"
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "workspaceId",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "List Projects"
      },
      "post": {
        "operationId": "createProject",
        "tags": [
          "Projects"
        ],
        "description": "Create a new project in a workspace",
        "responses": {
          "200": {
            "description": "Project created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "workspaceId": {
                      "type": "string"
                    },
                    "slug": {
                      "type": "string"
                    },
                    "icon": {
                      "type": "string",
                      "nullable": true
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "createdAt": {},
                    "isPublic": {
                      "type": "boolean",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "workspaceId",
                    "slug",
                    "icon",
                    "name",
                    "description",
                    "createdAt",
                    "isPublic"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "workspaceId": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "workspaceId",
                  "icon",
                  "slug"
                ]
              }
            }
          }
        },
        "summary": "Create Project"
      }
    },
...
    "/time-entry": {
      "post": {
        "operationId": "createTimeEntry",
        "tags": [
          "Time Entries"
        ],
        "description": "Create a new time entry for a task",
        "responses": {
          "200": {
            "description": "Time entry created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "taskId": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string",
                      "nullable": true
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "startTime": {},
                    "endTime": {},
                    "duration": {
                      "type": "number",
                      "nullable": true
                    },
                    "createdAt": {}
                  },
                  "required": [
                    "id",
                    "taskId",
                    "userId",
                    "description",
                    "startTime",
                    "duration",
                    "createdAt"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "taskId": {
                    "type": "string"
                  },
                  "startTime": {
                    "type": "string"
                  },
                  "endTime": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "required": [
                  "taskId",
                  "startTime"
                ]
              }
            }
          }
        },
        "summary": "Create Time Entry"
      }
    },
...
    "/label": {
      "post": {
        "operationId": "createLabel",
        "tags": [
          "Labels"
        ],
        "description": "Create a new label in a workspace",
        "responses": {
          "200": {
            "description": "Label created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "color": {
                      "type": "string"
                    },
                    "createdAt": {},
                    "taskId": {
                      "type": "string",
                      "nullable": true
                    },
                    "workspaceId": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "color",
                    "createdAt",
                    "taskId",
                    "workspaceId"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "color": {
                    "type": "string"
                  },
                  "workspaceId": {
                    "type": "string"
                  },
                  "taskId": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "color",
                  "workspaceId"
                ]
              }
            }
          }
        },
        "summary": "Create Label"
      }
    },
...
    "/notification": {
      "get": {
        "operationId": "listNotifications",
        "tags": [
          "Notifications"
        ],
        "description": "Get all notifications for the current user",
        "responses": {
          "200": {
            "description": "List of notifications",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "userId": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "content": {
                        "type": "string",
                        "nullable": true
                      },
                      "type": {
                        "enum": [
                          "info",
                          "task_created",
                          "workspace_created",
                          "task_status_changed",
                          "task_assignee_changed",
                          "time_entry_created"
                        ]
                      },
                      "isRead": {
                        "type": "boolean"
                      },
                      "resourceId": {
                        "type": "string"
                      },
                      "resourceType": {
                        "enum": [
                          "task",
                          "workspace"
                        ]
                      },
                      "createdAt": {}
                    },
                    "required": [
                      "id",
                      "userId",
                      "title",
                      "content",
                      "type",
                      "createdAt"
                    ]
                  }
                }
              }
            }
          }
        },
        "summary": "List Notifications"
      },
      "post": {
        "operationId": "createNotification",
        "tags": [
          "Notifications"
        ],
        "description": "Create a new notification for a user",
        "responses": {
          "200": {
            "description": "Notification created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string",
                      "nullable": true
                    },
                    "type": {
                      "enum": [
                        "info",
                        "task_created",
                        "workspace_created",
                        "task_status_changed",
                        "task_assignee_changed",
                        "time_entry_created"
                      ]
                    },
                    "isRead": {
                      "type": "boolean"
                    },
                    "resourceId": {
                      "type": "string"
                    },
                    "resourceType": {
                      "enum": [
                        "task",
                        "workspace"
                      ]
                    },
                    "createdAt": {}
                  },
                  "required": [
                    "id",
                    "userId",
                    "title",
                    "content",
                    "type",
                    "createdAt"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "relatedEntityId": {
                    "type": "string"
                  },
                  "relatedEntityType": {
                    "type": "string"
                  }
                },
                "required": [
                  "userId",
                  "title",
                  "message",
                  "type"
                ]
              }
            }
          }
        },
        "summary": "Create Notification"
      }
    },
...
    "/invitation/{id}": {
      "get": {
        "operationId": "getInvitationDetails",
        "tags": [
          "Invitations"
        ],
        "description": "Get details of a specific invitation by ID",
        "responses": {
          "200": {
            "description": "Invitation details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "invitation": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "workspaceName": {
                          "type": "string"
                        },
                        "inviterName": {
                          "type": "string"
                        },
                        "expiresAt": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "expired": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "workspaceName",
                        "inviterName",
                        "expiresAt",
                        "status",
                        "expired"
                      ]
                    },
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "valid"
                  ]
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "Get Invitation Details"
      }
    },
...
  },
...
}